requirements.txt
WPyrogram
tgcrypto
aiohttp[speedups]
jdatetime
configparser
peewee